% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.treatments.R
\name{add.treatments}
\alias{add.treatments}
\title{Add more descriptive treatments to the data set}
\usage{
add.treatments(D)
}
\arguments{
\item{D}{The dataset you want to add variables to}
}
\value{
The same dataset, but with new variables describing fertilization treatments and geomorphology
}
\description{
Add more descriptive treatments to the data set
}
\examples{

}
