% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cleaning.R
\name{data.cleaning}
\alias{data.cleaning}
\title{Clean variables identifying observations}
\usage{
data.cleaning(D)
}
\arguments{
\item{D}{The dataset you want to clean.}
}
\value{
The same dataset, but with uniformized Parcelle, Traitement and Exclos columns
}
\description{
Clean variables identifying observations
}
\examples{

}
