% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Environment_Date}
\alias{Environment_Date}
\title{Surface environment of plots}
\format{
A data frame with 542 rows and 22 variables:
\describe{
  \item{\code{Date}}{double Date at which the measurement has been performed}
  \item{\code{Parcelle}}{integer Name of the site}
  \item{\code{Traitement}}{character Name of the treatment. PA : Polygons slightly drained, PAS : Polygons drained by a secondary ice wedge,
  PS : polygons close to the shore, submitted to abundant eolian silt deposition, PO : flat to high-center mesic polygons,  PR : mesic meadow,
  1 : low-centered, wet polygon; 4 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3; 6 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3,
  10 : wet polygons fertilized yearly with 3 g.m-2 of H3PO4; 14 : wet polygons fertilized yearly with 1 g.m-2 of H3PO4 amd 5 g.m-2 of NH4NO3}
  \item{\code{Exclos}}{integer Fencing. Exclos : plots protected constantly by a fence; Temoin : unprotected plots}
  \item{\code{Thaw_depth}}{double Depth of the thaw front, as measured physically with a graduated probe, cm}
  \item{\code{Soil_temp}}{double Soil temperature at ~5cm, Celsius degree}
  \item{\code{SVWC}}{double Soil Volumetric Water content as derived from permittivity measurements, cm-3.cm-3}
  \item{\code{WaterTable_depth}}{double Depth of the water table, as measured in a hole, cm}
  \item{\code{pH}}{double Shallow soil pH, unitless}
  \item{\code{NDVI}}{double Normalized Difference Vegetation Index, 0-1}
  \item{\code{PRI}}{double Photochemical Reflectance index, -1 - 1}
  \item{\code{Reflectance_blue}}{double Reflectance in the blue color band, 0-1}
  \item{\code{Reflectance_green}}{double Reflectance in the green color band, 0-1}
  \item{\code{Reflectance_red}}{double Reflectance in the red color band, 0-1}
  \item{\code{Reflectance_visible}}{double Reflectance in the visible, 0-1}
  \item{\code{Reflectance_NIR}}{double Reflectance in the NIR color band, 0-1}
  \item{\code{Albedo}}{double Surface albedo, 0-1}
  \item{\code{Geomorpho}}{integer A formulation of the Traitement columns focusing on discrete
  habitat differences}
  \item{\code{Fertilization_N}}{integer Specifies the quantity of N added each year in g.m-2 of NH4NO3}
  \item{\code{Fertilization_P}}{integer Specifies the quantity of P added each year in g.m-2 of H3PO4}
  \item{\code{Fertilization}}{integer A formulation of the Traitement columns with a zero to each unfertilized plots}
  \item{\code{Fertilization_P_bin}}{integer Binary variable indicating if the plot receive annual P fertilization}
}
}
\usage{
Environment_Date
}
\description{
This dataset compiles measures of environmental describing the plots
}
\details{
DETAILS
}
\keyword{datasets}
