% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Soil_Physic_Horizons}
\alias{Soil_Physic_Horizons}
\title{Physical properties of soil horizons}
\format{
A data frame with 280 rows and 35 variables:
\describe{
  \item{\code{Date}}{character Date at which the sampled has been processed}
  \item{\code{Parcelle}}{integer Name of the site}
  \item{\code{Traitement}}{character Name of the treatment. PA : Polygons slightly drained, PAS : Polygons drained by a secondary ice wedge,
  PS : polygons close to the shore, submitted to abundant eolian silt deposition, PO : flat to high-center mesic polygons,  PR : mesic meadow,
  1 : low-centered, wet polygon; 4 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3; 6 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3,
  10 : wet polygons fertilized yearly with 3 g.m-2 of H3PO4; 14 : wet polygons fertilized yearly with 1 g.m-2 of H3PO4 amd 5 g.m-2 of NH4NO3}
  \item{\code{Exclos}}{integer Fencing. Exclos : plots protected constantly by a fence; Temoin : unprotected plots}
  \item{\code{Depth_up}}{double Upper depth of the horizon}
  \item{\code{Depth_down}}{double Lower depth of the horizon}
  \item{\code{Depth_mean}}{double Mean depth of the horizon}
  \item{\code{Oxydo_Reduction}}{double Did the horizon displayed trace of oxydo-reduction?}
  \item{\code{Volume}}{double Volume of the horizon, cm-3}
  \item{\code{Density}}{double Bulk density of the horizon, g.cm-3}
  \item{\code{LOI}}{double Proportion of organic matter, as lost on ignition, (0-1)}
  \item{\code{V_om}}{double Volumic fraction of organic matter}
  \item{\code{V_oms}}{double Proportion of the volume of the solide attributed to organic matter}
  \item{\code{Mineral}}{double Proportion of mineral material (1-LOI, 0-1)}
  \item{\code{V_min}}{double Volumic fraction of mineral material}
  \item{\code{V_mins}}{double Proportion of the volume of the solide attributed to minerals}
  \item{\code{Particle_Density}}{double Particle density as approximated by weighting a known volume of powder, g.cm-3}
  \item{\code{Particle_Density_computed}}{double Particle density computed following Balland and Arp (2005), g.cm-3}
  \item{\code{Porosity}}{double Porosity computed using approximation of particle density (0-1)}
  \item{\code{Porosity_computed}}{double Porosity computed using computed particle density (0-1)}
  \item{\code{GWC}}{double Gravimetric Water Content, g.g-1}
  \item{\code{VWC}}{double Volumetric Water Content, g.cm-3}
  \item{\code{theta_sat}}{double Proportion of the porosity filled with water}
  \item{\code{K_solid}}{double Thermal conductivity of solids}
  \item{\code{K_dry}}{double Thermal conductivity of dry materials}
  \item{\code{K_sat}}{double Thermal conductivity of the saturated soil}
  \item{\code{K_e}}{double Impact of water content on soil thermal conductivity}
  \item{\code{K_soil}}{double Raw thermal conductivity of soil}
  \item{\code{VolHeatCap}}{double Volumetric heat capacity of soil, as a weighted mean of its components}
  \item{\code{ThermDiff}}{double Thermal diffusivity of soil}
  \item{\code{Geomorpho}}{integer A formulation of the Traitement columns focusing on discrete
  habitat differences}
  \item{\code{Fertilization_N}}{integer Specifies the quantity of N added each year in g.m-2 of NH4NO3}
  \item{\code{Fertilization_P}}{integer Specifies the quantity of P added each year in g.m-2 of H3PO4}
  \item{\code{Fertilization}}{integer A formulation of the Traitement columns with a zero to each unfertilized plots}
  \item{\code{Fertilization_P_bin}}{integer Binary variable indicating if the plot receive annual P fertilization}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Soil_Physic_Horizons
}
\description{
This dataset describes the physical properties of each homogeneous horizon
}
\keyword{datasets}
