% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Vascular_Traits_Plot}
\alias{Vascular_Traits_Plot}
\title{Traits of Vascular Plants in each plot}
\format{
A data frame with 340 rows and 22 variables:
\describe{
  \item{\code{Parcelle}}{integer Name of the site}
  \item{\code{Traitement}}{character Name of the treatment. PA : Polygons slightly drained, PAS : Polygons drained by a secondary ice wedge,
  PS : polygons close to the shore, submitted to abundant eolian silt deposition, PO : flat to high-center mesic polygons,  PR : mesic meadow,
  1 : low-centered, wet polygon; 4 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3; 6 : wet polygons fertilized yearly with 1 g.m-2 of NH4NO3,
  10 : wet polygons fertilized yearly with 3 g.m-2 of H3PO4; 14 : wet polygons fertilized yearly with 1 g.m-2 of H3PO4 amd 5 g.m-2 of NH4NO3}
  \item{\code{Exclos}}{integer Exclos : plots protected constantly by a fence; Temoin : unprotected plots}
  \item{\code{Herbivorie}}{character Grazing (reformulation of Exclos). Ungrazed : plots protected constantly by a fence; Grazed : unprotected plots}
  \item{\code{Sp}}{character Code of species}
  \item{\code{Leaf_N15}}{double Leaf d15N, in per mill}
  \item{\code{LNC}}{double Leaf Nitrogen Content, in percent}
  \item{\code{Leaf_C13}}{double Leaf d13C, in per mill}
  \item{\code{LCC}}{double Leaf Carbon Content, in percent}
  \item{\code{Root_N15}}{double Root d15N, in per mill}
  \item{\code{RNC}}{double Root Nitrogen Content, in percent}
  \item{\code{Leaf_pH}}{double Leaf Hydrogen potential}
  \item{\code{Hveg}}{double Vegetative height in cm, measured as the distance between the base of the plant
  and the extremity of the extended leaves. Measured on height individuals per plot or less}
  \item{\code{LDMC}}{double Leaf Dry Matter Content, in g_dry.g_fresh^-1}
  \item{\code{RDMC}}{double Root Dry Matter Content, in g_dry.g_fresh^-1}
  \item{\code{LA}}{double Leaf Area, in cm^2}
  \item{\code{SLA}}{double Specific Leaf Area, in cm^2.g^-1}
  \item{\code{Geomorpho}}{integer A formulation of the Traitement columns focusing on discrete
  habitat differences}
  \item{\code{Fertilization_N}}{integer Specifies the quantity of N added each year in g.m-2 of NH4NO3}
  \item{\code{Fertilization_P}}{integer Specifies the quantity of P added each year in g.m-2 of H3PO4}
  \item{\code{Fertilization}}{integer A formulation of the Traitement columns with a zero to each unfertilized plots}
  \item{\code{Fertilization_P_bin}}{integer Binary variable indicating if the plot receive annual P fertilization}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
Vascular_Traits_Plot
}
\description{
This dataset contains the mean traits of each species in each plot,
measured either on 10-12 individuals or on every individuals pulled together, depending on their abundances.
}
\keyword{datasets}
